requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==39.0.2
daphne==4.0.0
Django==4.1.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
exceptiongroup==1.1.0
fontawesomefree==6.2.1
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
outcome==1.2.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
PySocks==1.7.1
redis==4.5.1
selenium==4.8.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
wsproto==1.2.0
zope.interface==5.5.2
